refactor: Update import paths for handle-deposit-button and span-button-handler
