+++ src/handlers/span-button-handler.ts
@@ -1,22 +1,8 @@
-import { createLogger } from '../logger';
+import { createLogger } from '../utils/logger';
 
 const logger = createLogger();
 
 export const handleButtonWithSpan = (spanText: string): void => {
-	// const autoRegisterCheckbox = document.getElementById(
-	//   "autoRegister"
-	// ) as HTMLInputElement;
-
-	// if (!autoRegisterCheckbox) {
-	//   console.error("Auto register checkbox not found");
-	//   return;
-	// }
-
-	// if (!autoRegisterCheckbox.checked) {
-	//   console.log("Auto register checkbox is not checked");
-	//   return;
-	// }
-
 	const spans = document.querySelectorAll('button span');
 	let buttonClicked = false;
